Remove container has check for plugins
Symfony container performs the check itself but seems to use another mechanism.
Anyway, the removed check fails for a plugin defined in a servies.yml. Without the
manual check loading the plugin works. A wrong service id causes symfony container to
fail with a similar message (dependency missing) so we can safely remove the check.